CBSE-6945: fix: show appropriate message when db not open (#2524)
* query when the db is not open case handled.
* corrected the replicator test variable using Int instead if the NSUInteger

+++ Objective-C/Tests/ReplicatorTest+Main.m
@@ -241,7 +241,7 @@
         __weak typeof(self) wSelf = self;
         id token = [r addChangeListener: ^(CBLReplicatorChange *change) {
             [wSelf verifyChange: change errorCode: 0 errorDomain: nil];
-            int whenValue = [when intValue];
+            NSUInteger whenValue = [when intValue];
             if (change.status.activity == whenValue) {
                 NSLog(@"****** Stop Replicator (when %@) ******", activities[whenValue]);
                 [change.replicator stop];
